+++ lib/src/unit_extension.dart
@@ -1,52 +1,60 @@
 part of dart_unitz;
 
-/// An extension to provide operators
+/// An extension to provide operators on Units
 ///
 ///
 extension UnitExtension on Unit {
   Unit operator +(Unit other) {
-    final Unit b = fromBase(other.toBase());
-    return newInstance(value + b.value);
+    final num b = fromBase(other.toBase());
+    return newInstance(value + b);
   }
 
   Unit operator -(Unit other) {
-    final Unit b = fromBase(other.toBase());
-    return newInstance(value - b.value);
+    final num b = fromBase(other.toBase());
+    return newInstance(value - b);
   }
 
   Unit operator *(Unit other) {
-    final Unit b = fromBase(other.toBase());
-    return newInstance(value * b.value);
+    final num b = fromBase(other.toBase());
+    return newInstance(value * b);
   }
 
   Unit operator /(Unit other) {
-    final Unit b = fromBase(other.toBase());
-    return newInstance(value / b.value);
+    final num b = fromBase(other.toBase());
+    return newInstance(value / b);
   }
 
   Unit operator ~/(Unit other) {
-    final Unit b = fromBase(other.toBase());
-    return newInstance(value ~/ b.value);
+    final num b = fromBase(other.toBase());
+    return newInstance(value ~/ b);
   }
 
   Unit operator %(Unit other) {
-    final Unit b = fromBase(other.toBase());
-    return newInstance(value % b.value);
+    final num b = fromBase(other.toBase());
+    return newInstance(value % b);
   }
 
   bool operator <(Unit other) {
-    return toBase().value < other.toBase().value;
+    final num a = base.newInstance(toBase()).formatted;
+    final num b = other.base.newInstance(other.toBase()).formatted;
+    return a < b;
   }
 
   bool operator <=(Unit other) {
-    return toBase().value <= other.toBase().value;
+    final num a = base.newInstance(toBase()).formatted;
+    final num b = other.base.newInstance(other.toBase()).formatted;
+    return a <= b;
   }
 
   bool operator >(Unit other) {
-    return toBase().value > other.toBase().value;
+    final num a = base.newInstance(toBase()).formatted;
+    final num b = other.base.newInstance(other.toBase()).formatted;
+    return a > b;
   }
 
   bool operator >=(Unit other) {
-    return toBase().value >= other.toBase().value;
+    final num a = base.newInstance(toBase()).formatted;
+    final num b = other.base.newInstance(other.toBase()).formatted;
+    return a >= b;
   }
 }