+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Dart Unitz
 
+## 0.0.3
+
+- Simplify base conversion requirements.
+- Equality now accounts for the 'formatted' value.
+
 ## 0.0.2
 
 - Allow types to be passed instead of instances.